+++ client-addon/src/components/ApolloView.vue
@@ -1,0 +1,9 @@
+<template>
+  <div class="apollo-view">
+    <ContentView
+      title="Apollo GraphQL"
+    >
+      Apollo Engine
+    </ContentView>
+  </div>
+</template>
