+++ client-addon/src/components/widgets/ApolloEngine.vue
@@ -1,0 +1,5 @@
+<template>
+  <div class="apollo-engine">
+    Engine
+  </div>
+</template>
