+++ ui/views.js
@@ -1,0 +1,7 @@
+module.exports = api => {
+  api.addView({
+    id: 'org.akryum.vue-apollo.apollo',
+    name: 'org.akryum.vue-apollo.routes.apollo',
+    tooltip: 'Apollo GraphQL',
+  })
+}
